+++ perfkitbenchmarker/resource.py
@@ -27,13 +27,19 @@ from perfkitbenchmarker import vm_util
 
 
 class BaseResource(object):
-  """An object representing a cloud resource."""
+  """An object representing a cloud resource.
+
+  Attributes:
+    created: True if the resource has been created.
+    pkb_managed: Whether the resource is managed (created and deleted) by PKB.
+  """
 
   __metaclass__ = abc.ABCMeta
 
-  def __init__(self):
+  def __init__(self, user_managed=False):
     super(BaseResource, self).__init__()
-    self.created = False
+    self.created = user_managed
+    self.user_managed = user_managed
 
     # Creation and deletion time information
     # that we may make use of later.
@@ -65,6 +71,18 @@ class BaseResource(object):
     exceptions.
     """
     raise NotImplementedError()
+
+  def _WaitUntilReady(self):
+    """Return true if the underlying resource is ready.
+
+    Supplying this method is optional.  Use it when a resource can exist
+    without being ready.  If the subclass does not implement
+    it then it just returns true.
+
+    Returns:
+      True if the resource was ready in time, False if the wait timed out.
+    """
+    return True
 
   def _PostCreate(self):
     """Method that will be called once after _CreateReource is called.
@@ -126,11 +144,18 @@ class BaseResource(object):
 
   def Create(self):
     """Creates a resource and its dependencies."""
+    if self.user_managed:
+      return
     self._CreateDependencies()
     self._CreateResource()
+    ready = self._WaitUntilReady()
+    if not ready:
+      raise Exception('Wait for resource to be read timed out, giving up')
     self._PostCreate()
 
   def Delete(self):
     """Deletes a resource and its dependencies."""
+    if self.user_managed:
+      return
     self._DeleteResource()
     self._DeleteDependencies()
